start of refactor

+++ src/Servers/Kestrel/shared/test/CompositeKestrelTrace.cs
@@ -241,7 +241,8 @@ namespace Microsoft.AspNetCore.Testing
             _trace1.InvalidResponseHeaderRemoved();
             _trace2.InvalidResponseHeaderRemoved();
         }
-        public void Http3ConnectionClosed(string connectionId, int highestOpenedStreamId)
+        
+        public void Http3ConnectionClosed(string connectionId, long highestOpenedStreamId)
         {
             _trace1.Http3ConnectionClosed(connectionId, highestOpenedStreamId);
             _trace2.Http3ConnectionClosed(connectionId, highestOpenedStreamId);
